+++ packages/expo-config-plugin/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2026-06-12
+
+### Changed
+- Bumped @rnrepo/build-tools dependency to 0.1.5 (#359)
+
+## [0.3.1] - 2026-05-21
+
+### Changed
+- Bumped @rnrepo/build-tools dependency to 0.1.4 (#344)
+
 ## [0.3.0-beta.3] - 2026-04-30
 
 ### Changed